requirements.txt
asgiref==3.8.1
Django==5.1
djangorestframework==3.14.0
pytz==2023.3.post1
sqlparse==0.4.4
gunicorn==21.2.0
psycopg2-binary==2.9.6
djangorestframework-simplejwt==5.3.1
django-lifecycle==1.2.4
drf-access-policy==1.5.0
django-filter==25.1
drf-nested-routers==0.94.1